backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai
